moved the toHuman function to StringUtil-Class
