+++ src/rappsilber/applications/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * main xi-search classes and command-line executables. e.g tools + the various 
+ * incarnations main xi-class that currently exists
+ */
+package rappsilber.applications;
